// front/lib/actions/types/retrieval.ts
export type DataSourceFilter = {
  parents: { in: string[]; not: string[] } | null;
  tags?: TagsFilter;
};

export type DataSourceConfiguration = {
  workspaceId: string;
  dataSourceViewId: string;
  filter: DataSourceFilter;
};
// ...
export type RetrievalQuery = "auto" | "none";
export type RetrievalTimeframe = "auto" | "none" | TimeFrame;
export type RetrievalConfigurationType = {
  id: ModelId;
  sId: string;

  type: "retrieval_configuration";

  dataSources: DataSourceConfiguration[];
  query: RetrievalQuery;
  relativeTimeFrame: RetrievalTimeframe;
  topK: number | "auto";

  name: string;
  description: string | null;
};

/**
 * Retrieval action
 */

export interface RetrievalDocumentChunkType {
  offset: number;
  score: number | null;
  text: string;
}

export interface RetrievalDocumentType {
  chunks: RetrievalDocumentChunkType[];
  documentId: string;
  dataSourceView: DataSourceViewType | null;
  id: ModelId;
  reference: string; // Short random string so that the model can refer to the document.
  score: number | null;
  sourceUrl: string | null;
  tags: string[];
  timestamp: number;
}
// ...
export interface RetrievalActionType extends BaseAction {
  id: ModelId; // AgentRetrievalAction
  agentMessageId: ModelId; // AgentMessage

  params: {
    relativeTimeFrame: TimeFrame | null;
    query: string | null;
    topK: number;
    tagsIn: string[] | null;
    tagsNot: string[] | null;
  };
  functionCallId: string | null;
  functionCallName: string | null;
  documents: RetrievalDocumentType[] | null;
  step: number;
  type: "retrieval_action";
}